% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_icd10cm_CHAT_2023}
\alias{icd_icd10cm_CHAT_2023}
\title{DOH CHAT ICD-10-CM & CSS}
\format{
a data.table with 71,807 rows and eight columns: \code{icd10_desc},
\code{icd10}, \code{ccs_lvl_1_desc}, \code{ccs_lvl_1}, \code{ccs_lvl_2_desc},
\code{ccs_lvl_2}, \code{ccs_lvl_3_desc}, & \code{ccs_lvl_3}.
}
\source{
https://secureaccess.wa.gov/ >> Community Health Assessment Tool (CHAT)
>> User Guides >> 'CHAT Hospital CCS Codes Sorted By ICD10cm Codes' on Jan 10, 2023
}
\usage{
icd_icd10cm_CHAT_2023
}
\description{
DOH CHAT ICD-10-CM & CSS with descriptions and levels of CCS.

This can be merged/joined with the contents of
\code{rads.data::icd_icd10cm_groups} to get the corresponding
\code{cause.category} and \code{leading.cause} of hospitalization
group identifiers. \strong{Please note} that the complete list
of leading causes of hospitalization should include injury related
hospitalizations by intent (\emph{hint, use \code{rads::chars_injury_matrix_count()}}).
}
\note{
CCS ==  Clinical Classifications Software (CCS)
}
\keyword{datasets}
