% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_icd10cm_codes_2022}
\alias{icd_icd10cm_codes_2022}
\title{CMS 2022 ICD-10-CM}
\format{
a data.table with 95,524 rows and five columns: \code{order}, \code{icd10}, \code{header}, \code{short}, \code{long}.
}
\source{
\url{https://www.cms.gov/files/zip/2022-code-descriptions-tabular-order-updated-02012022.zip}, found on
\url{https://www.cms.gov/medicare/icd-10/2022-icd-10-cm}
}
\usage{
icd_icd10cm_codes_2022
}
\description{
CMS (Centers for Medicare & Medicaid Services) 2022 ICD-10-CM
}
\note{
These are diagnosis codes, not procedure codes.

\code{header} == 0 == not valid for HIPAA-covered transactions.
}
\references{
\code{icd10-Order-Files-2022.pdf} in \url{https://www.cms.gov/files/zip/2022-code-descriptions-tabular-order-updated-02012022.zip}
}
\keyword{datasets}
