% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_icd10cm_groups}
\alias{icd_icd10cm_groups}
\title{Available groupings for ICD-10-CM causes of hospitalization}
\format{
a data.table with 283 rows and seven columns:
\code{leading.cause.group.num}, \code{leading.cause.group},
\code{leading.cause.group.alias}, \code{cause.category},
\code{ccs_lvl_1_desc}, \code{ccs_lvl_2_desc}, & \code{ccs_lvl_3_desc}.
}
\source{
Prepared by Danny Colombara, based on a draft approved by Eli Kern
on February 28, 2023. This replaces any and all other references for leading
causes of hospitalization reported by APDE.
}
\usage{
icd_icd10cm_groups
}
\description{
Provides a crosswalk between CCS Levels 1, 2, and 3 and 186
leading causes of hospitalization as well as the \code{cause.category}.

This would typically be merged with
\code{rads.data::icd_icd10cm_CHAT_2023} or the results of
\code{rads::chars_icd_ccs_count()}.
}
\note{
\strong{Don't forget!!!} These leading causes of hospitalization are
among non-injury hospitalizations. The complete list of leading causes of
hospitalizations should include injury hospitalizations by intent.
}
\keyword{datasets}
