% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_icd10pcs_codes_2022}
\alias{icd_icd10pcs_codes_2022}
\title{CMS 2022 ICD-10-PCS}
\format{
a data.table with 79,124 rows and five columns: \code{order}, \code{icd10}, \code{header}, \code{short}, \code{long}.
}
\source{
\code{icd10pcs_order_2022.txt} in \url{https://www.cms.gov/files/zip/errata-january-12-2022.zip}
}
\usage{
icd_icd10pcs_codes_2022
}
\description{
CMS (Centers for Medicare & Medicaid Services) 2022 ICD-10-PCS
}
\note{
These are procedure codes, not diagnosis codes.

\code{header} == 0 == not valid for HIPAA-covered transactions.
}
\references{
\code{icd10pcsOrderFile.pdf} in \url{https://www.cms.gov/files/zip/2022-icd-10-pcs-codes-file-updated-december-1-2021.zip}
}
\keyword{datasets}
