% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_nchs130causes}
\alias{icd_nchs130causes}
\title{NCHS 130 Selected Causes of Infant Death (long)}
\format{
A data.table with 18,936 rows and four variables: \code{causeid}, \code{cause.of.death}, \code{orig.coding},
\code{icd10}. Each row maps one ICD-10 code to one of the 130 causes of death.
}
\source{
Downloaded from
\url{https://secureaccess.wa.gov/doh/chat/Content/FilesForDownload/TechnicalNotes.pdf}
on 2/15/2023. It should represent table in
\href{https://www.cdc.gov/nchs/data/dvs/Part9InstructionManual2020-508.pdf}{ICD-10 Cause-of-Death Lists for Tabulating Mortality  Statistics}, on pages 18-20.
}
\usage{
icd_nchs130causes
}
\description{
NCHS 130 Selected Causes of Infant Death (long).
}
\details{
Tidied long version of \code{icd_nchs130causes_raw}. Converted by \verb{/data-raw/icd_nchs130causes.R}.
}
\references{
\itemize{
\item \href{https://github.com/PHSKC-APDE/rads.data/wiki/ICD-10}{A Brief Introduction to ICD-10 Death Codes}
\item \href{https://www.cdc.gov/nchs/data/dvs/Part9InstructionManual2020-508.pdf}{ICD-10 Cause-of-Death Lists for Tabulating Mortality  Statistics}
}
}
\keyword{datasets}
