% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_other_causes_of_death}
\alias{icd_other_causes_of_death}
\title{Other Causes of Death (long)}
\format{
A data.table with 3,063 rows and four variables: \code{cause.of.death},
\code{orig.coding}, \code{icd10}, and \code{source}. Each row maps one
specific ICD 10 code to a given cause of death.
}
\usage{
icd_other_causes_of_death
}
\description{
This table contains ICD 10 definitions of causes of death that are of
interest for systematic analyses, but are not included in the CDC 113
Causes of Death (\link{icd_nchs113causes}) or the injury
matrix (\link{icd10_death_injury_matrix}). It includes causes such
as carbon monoxide deaths, drug related deaths, heat stress deaths, etc.
}
\note{
Tidied long version of \verb{/data-raw/icd_other_causes_of_death_raw} PLUS
an Excel workbook downloaded from DOH/CHAT and saved to SharePoint.
Created by \verb{/data-raw/icd_other_causes_of_death.R}.
}
\references{
\itemize{
\item \href{https://github.com/PHSKC-APDE/rads.data/wiki/ICD-10}{A Brief Introduction to ICD-10 Death Codes}
\item \href{https://secureaccess.wa.gov/doh/chat/Entry.mvc}{WA DOH CHAT (User Guides >> Definition of "Special Code Sets")}
\item \href{https://www.cdc.gov/nchs/data/nvsr/nvsr70/nvsr70-08-508.pdf}{National Vital Statistics Reports, Vol 70, No. 8}
}
}
\keyword{datasets}
