% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_poverty_groups}
\alias{misc_poverty_groups}
\title{Poverty groupings based on 200 percent FPL}
\format{
a data.table with 578 rows and six columns: \code{geo_type}, \code{geo_id}, \code{census_vintage}, \code{pov200grp},
\code{source}, \code{creation_date}.
}
\source{
ACS 2017-2021 tabular estimates processed by \url{https://github.com/PHSKC-APDE/chi/blob/main/acs/2021_2017/03_calculations_5_year_tabular_data.R}
with results saved in "//dphcifs/APDE-CDIP/ACS/2021_2017_5_year/Analysis/03_ACS_Calculations/output_ACS_data_2021_2017_by_geography.xlsx"
}
\usage{
misc_poverty_groups
}
\description{
Tract and ZCTA level poverty groupings based on the proportion of the population below 200 percent of the Federal Poverty Level (FPL)
}
\note{
These replace the previous poverty groups (low, medium, high) that were based on 100 percent FPL from 2008-2012 ACS data.
}
\references{
Code based on \url{https://github.com/PHSKC-APDE/pers_dvc/blob/main/R/2020_poverty_groupings_v2.qmd}
}
\keyword{datasets}
