% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation_soc_2018_definitions}
\alias{occupation_soc_2018_definitions}
\title{2018 BLS Standard Occupational Classification (SOC) definitions.}
\format{
A data.table with 1447 rows and five variables: \code{'SOC.Group'},
\code{'SOC.Code'}, \code{'SOC.Title'}, \code{'SOC.Definition'}, \code{'uploaded'}.
}
\source{
\url{https://www.bls.gov/soc/soc_2018_definitions.xls}
}
\usage{
occupation_soc_2018_definitions
}
\description{
2018 BLS Standard Occupational Classification (SOC) definitions.
}
\details{
Detailed definitions for each 2018 BLS SOC code / title
}
\keyword{datasets}
