% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation_soc_2018_essential_workers_key}
\alias{occupation_soc_2018_essential_workers_key}
\title{'Essential Worker' definitions using 2018 BLS Standard Occupational Classification (SOC) codes.}
\format{
A data.table with 1447 rows and five variables: \code{essential}, \code{group},
  \code{code}, \code{title}, \code{notes}.
}
\source{
\url{https://www.lmiontheweb.org/more-than-half-of-u-s-workers-in-critical-occupations-in-the-fight-against-covid-19/}
}
\usage{
occupation_soc_2018_essential_workers_key
}
\description{
'Essential Worker' definitions using 2018 BLS Standard Occupational Classification (SOC) codes.
    One of many essential worker definitions circulating early in the COVID-19 pandemic.
}
\keyword{datasets}
