% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation_soc_structure_2010}
\alias{occupation_soc_structure_2010}
\title{2010 BLS Standard Occupational Classification (SOC) structure.}
\format{
A data.table with 1421 rows and twelve variables: \code{group}, \code{code},
  \code{major}, \code{minor}, \code{broad}, \code{detailed}, \code{title},
  \code{major.title}, \code{minor.title}, \code{broad.title},
  \code{detailed.title}, \code{notes}.
}
\source{
\url{https://www.bls.gov/soc/2010/2010_major_groups.htm}
}
\usage{
occupation_soc_structure_2010
}
\description{
2010 BLS Standard Occupational Classification (SOC) structure.
    SOC codes with hierarchy information.
}
\keyword{datasets}
