% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation_soc_structure_2018}
\alias{occupation_soc_structure_2018}
\title{2018 BLS Standard Occupational Classification (SOC) structure.}
\format{
A data.table with 1447 rows and eleven variables: \code{group}, \code{code},
  \code{major}, \code{minor}, \code{broad}, \code{detailed}, \code{title},
  \code{major.title}, \code{minor.title}, \code{broad.title}, \code{detailed.title}.
}
\source{
\url{https://www.bls.gov/soc/2018/2018_major_groups.htm}
}
\usage{
occupation_soc_structure_2018
}
\description{
Detailed definitions for each 2018 BLS SOC code / title
}
\keyword{datasets}
