% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_hra_pop_18}
\alias{population_hra_pop_18}
\title{2018 Population by HRA.}
\format{
A data.table with 48 rows and eight variables: \code{year}, \code{hra_id},
  \code{hra_name}, \code{age}, \code{gender}, \code{race}, \code{ethnicity},
  \code{population}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
population_hra_pop_18
}
\description{
Health Reporting Area (HRA) 2018 population. Not clear why this was worth saving in GitHub.
Not stratified by demographics.
}
\keyword{datasets}
