% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_reference_pop_11_age_groups}
\alias{population_reference_pop_11_age_groups}
\title{Standard reference populations with 11 age categories.}
\format{
A data.table with 11 rows and 7 variables: \code{standard}, \code{agecat},
  \code{age_start}, \code{age_end}, \code{pop}, \code{source}, \code{uploaded}.
}
\source{
\url{https://www.cdc.gov/nchs/data/statnt/statnt20.pdf}
}
\usage{
population_reference_pop_11_age_groups
}
\description{
One standard reference populations with 11 non-overlapping age categories.
2000 US Standard Population with 11 age categories is used by WA DOH CHAT
as of August 26, 2022.
}
\keyword{datasets}
