% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_reference_pop_19_age_groups}
\alias{population_reference_pop_19_age_groups}
\title{Standard reference populations with 19 age categories.}
\format{
A data.table with 285 rows and 7 variables: \code{standard}, \code{agecat},
  \code{age_start}, \code{age_end}, \code{pop}, \code{source}, \code{uploaded}.
}
\source{
\url{https://seer.cancer.gov/stdpopulations}
}
\usage{
population_reference_pop_19_age_groups
}
\description{
Fifteen standard reference populations with 19 non-overlapping age categories.
}
\keyword{datasets}
