% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_reference_pop_single_age_to_84}
\alias{population_reference_pop_single_age_to_84}
\title{Standard population for single ages through 84.}
\format{
A data.table with 258 rows and 6 variables: \code{standard}, \code{agecat},
  \code{age_start}, \code{age_end}, \code{pop}, \code{source}.
}
\source{
\url{https://seer.cancer.gov/stdpopulations}
}
\usage{
population_reference_pop_single_age_to_84
}
\description{
Standard population for single ages through 84.
After 84, binned into 85-120. Three standards.
}
\keyword{datasets}
