% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_reference_pop_single_age_to_99}
\alias{population_reference_pop_single_age_to_99}
\title{Standard population for single ages through 99.}
\format{
A data.table with 202 rows and 6 variables: \code{standard}, \code{agecat},
  \code{age_start}, \code{age_end}, \code{pop}, \code{source}.
}
\source{
\url{https://seer.cancer.gov/stdpopulations}
}
\usage{
population_reference_pop_single_age_to_99
}
\description{
Standard population for single ages through 99.
After 99, binned into 99-120. Two standards.
}
\keyword{datasets}
