% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_wapop_codebook_values}
\alias{population_wapop_codebook_values}
\title{Codebook for WA OFM population demographics.}
\format{
A data.table with 117 rows and 5 variables: \code{varname}, \code{code},
  \code{code_label}, \code{code_label_long}, \code{code_label_long2}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
population_wapop_codebook_values
}
\description{
Codebook for WA OFM population demographics.
   Translation of OFM population variables into standard demographic terms.
   Most likely to be used with raw data in SQL.
}
\keyword{datasets}
