% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_acs_council_districts}
\alias{spatial_acs_council_districts}
\title{Crosswalk from Census tracts to political districts.}
\format{
A data.table with 397 rows and 6 variables: \code{geo_id}, \code{name},
  \code{tract}, \code{kccd}, \code{sea_ccd}, , \code{uploaded}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_acs_council_districts
}
\description{
Crosswalk from Census tracts to political districts.
Census/ACS tracts mapped to King County and Seattle districts.
}
\keyword{datasets}
