% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_acs_hra_region_place_etc}
\alias{spatial_acs_hra_region_place_etc}
\title{Crosswalk hra > region > coo place > geoid > place > tract > block group}
\format{
A data.table with 276 rows and 9 variables: \code{hraname}, \code{region},
  \code{coo.place}, \code{geoid}, \code{name}, \code{place}, \code{tract},
  \code{blkgrp}, \code{notes}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_acs_hra_region_place_etc
}
\description{
Crosswalk hra > region > coo place > geoid > place > tract > block group
}
\keyword{datasets}
