% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_block10_to_hra20_to_region20}
\alias{spatial_block10_to_hra20_to_region20}
\title{Spatial crosswalk for Census 2010 blocks to 2020 HRAs and 2020 regions}
\format{
A data.table with 35,830 rows and 6 variables:

\code{hra20_id}: Numeric id for the HRA,

\code{GEOID10}: 2010 Census block id,

\code{hra20_name}: Name of the HRA,

\code{region_id}: numeric id for the region,

\code{region_name}: Name of the region,

\code{creation_date}: Date the file was created
}
\source{
2010 Census blocks (\verb{//dphcifs/APDE-CDIP/Shapefiles/Census_2010/block/}) intersected
with the 2020 HRA shapefile, with region labels merged in from \code{spatial_block20_to_hra20_to_region20}.
}
\usage{
spatial_block10_to_hra20_to_region20
}
\description{
Spatial crosswalk for Census 2010 blocks to 2020 HRAs and 2020 regions
}
\keyword{datasets}
