% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_blocks00_to_hra}
\alias{spatial_blocks00_to_hra}
\title{Census 2000 crosswalk Block > Block Group > Tract > HRA.}
\format{
A data.table with 24,475 rows and 4 variables: \code{btfid = block}, \code{gtfid = block group},
  \code{stfid = tract}, \code{hra}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_blocks00_to_hra
}
\description{
Census 2000 crosswalk Block > Block Group > Tract > HRA.
}
\keyword{datasets}
