% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_blocks10_to_hra}
\alias{spatial_blocks10_to_hra}
\title{Census 2010 crosswalk blocks >> hras}
\format{
A data.table with 35838 rows and 3 variables: \code{geo_id_blk}, \code{hra},
  \code{vid}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_blocks10_to_hra
}
\description{
Census 2010 crosswalk blocks >> hras
}
\keyword{datasets}
