% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_blocks10_to_hra_to_region}
\alias{spatial_blocks10_to_hra_to_region}
\title{Census 2010 crosswalk blocks >> hra >> region}
\format{
A data.table with 35,837 rows and 8 variables: \code{geo_id_blk}, \code{hra},
  \code{vid}, \code{region}, \code{region_id}, \code{region_id_old}, \code{svc},
  \code{svcid}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_blocks10_to_hra_to_region
}
\description{
Census 2010 crosswalk blocks >> hra >> region
}
\keyword{datasets}
