% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_county_codes}
\alias{spatial_county_codes}
\title{Crosswalk between 2010 county id, code, fips code and name}
\format{
A data.table with 39 rows and 7 variables: \code{geo_county_name},
\code{geo_county_code_fips }, \code{geo_county_fips_long},
\code{geo_county_code_order}, \code{geo_county_code_gnis}.
\code{geo_county_code_tiger}, \code{geo_county_code_aff}.
}
\source{
???
}
\usage{
spatial_county_codes
}
\description{
Crosswalk between 2010 county id, code, fips code and name
}
\keyword{datasets}
