% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_geocomp_blk10_kps}
\alias{spatial_geocomp_blk10_kps}
\title{Census 2010 ... most extensive crosswalk file ... the best}
\format{
A data.table with 61,692 rows and 51 variables:
\code{geo_id_blk10}, \code{geo_id_tra10}, \code{hra_id}, \code{hra_name},
\code{rgn_id}, \code{rgn_name}, \code{svc_id}, \code{svc_name},
\code{ccreg}, \code{rgn_ads}, \code{hracode_brfss}, \code{hracode_name},
\code{hracity_vid}, \code{hracity_id}, \code{hracity_ads},
\code{hracity_brfss}, \code{hracity_name}, \code{hracity2_id},
\code{hracity2_name}, \code{hrac_id}, \code{hrac_name}, \code{hrac2_id},
\code{hrac2_name}, \code{pvt_id}, \code{pvt_name}, \code{geo_id_scd10},
\code{scd_id}, \code{scd_name}, \code{scd_fipsco}, \code{scd_coname},
\code{scd}, \code{scd_ccddd}, \code{scd_esd}, \code{kc_scd},
\code{kc_scd_name}, \code{geo_id_cou10}, \code{cou_vid}, \code{cou_id},
\code{cou}, \code{wa_cou}, \code{fips_co}, \code{cou_name}, \code{kingco},
\code{tra_id}, \code{tra_name}, \code{leg_id}, \code{leg_name},
\code{con_id}, \code{con_name}, \code{scc_id}, \code{scc_name}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_geocomp_blk10_kps
}
\description{
Census 2010 ... APDE's most extensive crosswalk file.
Mike Smyser's extensive cross-walking file for King, Pierce, and Snomhomish
counties. It starts with 2010 census block ids and cross walks to HRAs,
KC regions, Seattle/Non-Seattle, Communities Count regions, school districts,
legislative districts, KC Council Districts and Seattle City Council
Districts.
}
\keyword{datasets}
