% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_hra20_to_bigcities}
\alias{spatial_hra20_to_bigcities}
\title{Spatial crosswalk between 2020 HRAs and King County's eight 'big cities'.}
\format{
A data.table with 34 rows and 3 variables:

  \code{hra20_id}: Numeric id for the HRA,

  \code{hra20_name}: Name of the HRA,

  \code{bigcity}: Name of the big city
}
\usage{
spatial_hra20_to_bigcities
}
\description{
A simple 1:1 crosswalk of which individual 2020 HRAs map to
King County's 'big cities'.
}
\keyword{datasets}
