% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_hra20_to_coo20}
\alias{spatial_hra20_to_coo20}
\title{Spatial crosswalk between 2020 HRAs and 2020 Communities of Opportunity areas}
\format{
A data.table with 19 rows and 3 variables:

  \code{hra20_id}: Numeric id for the HRA,

  \code{hra20_name}: Name of the HRA,

  \code{coo20_name}: numeric id for the COO area
}
\usage{
spatial_hra20_to_coo20
}
\description{
A simple crosswalk of which individual 2020 HRAs map to specific
COO areas.
}
\keyword{datasets}
