% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_hra20_to_region20}
\alias{spatial_hra20_to_region20}
\title{Spatial crosswalk between 2020 HRAs and 2020 regions.}
\format{
A data.table with 61 rows and 4 variables:

\code{hra20_id}: Numeric id for the HRA,

\code{hra20_name}: Name of the HRA,

\code{region_id}: numeric id for the region,

\code{region_name}: Name of the region
}
\source{
Derived from the unique HRA/region combinations in \code{spatial_block20_to_hra20_to_region20}.
}
\usage{
spatial_hra20_to_region20
}
\description{
A simple 1:1 crosswalk of which individual 2020 HRAs map to
specific regions.
}
\keyword{datasets}
