% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_hra_shortname_to_CHIname}
\alias{spatial_hra_shortname_to_CHIname}
\title{Crosswalk HRA short names to long names for CHI standards}
\format{
A data.table with 48 rows and 3 variables:
\code{cat1_group}, \code{cat1_group_alias},
\code{uploaded}.
}
\source{
\url{https://kc1.sharepoint.com/:x:/r/teams/DPH-CommunityHealthIndicators/}
}
\usage{
spatial_hra_shortname_to_CHIname
}
\description{
Crosswalk HRA short names to long names for CHI standards
}
\keyword{datasets}
