% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_hra_vid_region}
\alias{spatial_hra_vid_region}
\title{Crosswalk HRA's to regions}
\format{
A data.table with 48 rows and 4 variables:
\code{hra}, \code{vid}, \code{region}, \code{region_id}, \code{region_id_old}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_hra_vid_region
}
\description{
Crosswalk HRA's to regions
}
\keyword{datasets}
