% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_legislative_codes_to_names}
\alias{spatial_legislative_codes_to_names}
\title{Crosswalk between 2010 legislative district id, name, county, and url}
\format{
A data.table with 49 rows and 5 variables: \code{lgd_id}, \code{geo_year},
  \code{lgd_name}, \code{lgd_counties}, \code{lgd_url}.
}
\source{
kcitazrhpasqlprp16.azds.kingcounty.gov [ref].[pop_lgd_crosswalk]
}
\usage{
spatial_legislative_codes_to_names
}
\description{
Crosswalk between 2010 legislative district id, name, county, and url
}
\keyword{datasets}
