% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_school_codes_to_names}
\alias{spatial_school_codes_to_names}
\title{Crosswalk between 2010 school district id, name, state id, county code, and esd name}
\format{
A data.table with 295 rows and 3 variables: \code{geo_year}, \code{scd_id},
 and \code{scd_name}.
}
\source{
\url{https://secureaccess.wa.gov/doh/chat/Entry.mvc}
}
\usage{
spatial_school_codes_to_names
}
\description{
Crosswalk between 2010 school district id, name, state id, county code, and esd name
}
\keyword{datasets}
