% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_school_dist_to_region}
\alias{spatial_school_dist_to_region}
\title{Crosswalk between King County school districts and KC regions}
\format{
A data.table with 19 rows and 4 variables:
\code{school_district}, \code{geo_id}, \code{region}, \code{region_id}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_school_dist_to_region
}
\description{
Crosswalk between King County school districts and KC regions
}
\keyword{datasets}
