% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_tract10_to_hra20_pop}
\alias{spatial_tract10_to_hra20_pop}
\title{Spatial crosswalk between 2010 census tracts and 2020 HRAs. There is
one row between each tract and HRA that shares population.}
\format{
A data.table with 559 rows and 11 variables:

\code{GEOID10}: Census 2010 id,

\code{hra20_id}: HRA numeric id,

\code{s2t_fraction}: Percent of the population in tract that is also in hra20_id,

\code{isect_amount}: Number of people who live in the intersection of tract and hra20_id,

\code{tcoverage_amount}: Population of the hra20_id in any tract code,

\code{target_amount}: Total number of people within the HRA (including areas not covered by tracts),

\code{hra_name}: Name of the HRA,

\code{source_id}: tract,

\code{target_id}: hra20_id,

\code{method}: method to produce intersections,

\code{creation_date}: Date file created
}
\source{
2010 Census tracts intersected with the 2020 HRA shapefile using
\code{kcparcelpop::point_pop()} (point population method).
}
\usage{
spatial_tract10_to_hra20_pop
}
\description{
These are based on the point pop approach in \code{\link[spatagg:create_xwalk]{spatagg::create_xwalk()}} and broadly
follow that file structure.
}
\keyword{datasets}
