% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_tract10_to_hra20_region20_geog}
\alias{spatial_tract10_to_hra20_region20_geog}
\title{Spatial crosswalk between 2010 census tracts, 2020 HRAs, and 2020 regions. There is
one row between each tract and HRA that intersects geographically}
\format{
A data.table with 964 rows and 11 variables:
  \code{GEOID10}: tract id,
  \code{hra_id}: HRA numeric id,
  \code{s2t_fraction}: % of the area in ZIP that is also in hra_id,
  \code{isect_amount}: # of square survey feet in the intersection of tract and hra_id,
  \code{tcoverage_amount}: Area of the hra_id in any tract code,
  \code{target_amount}: Total area of  the HRA (including areas not covered by tracts)
  \code{hra_name}: Name of the HRA,
  \code{source_id}: tract,
  \code{target_id}: hra_id,
  \code{method}: method to produce intersections
  \code{creation_date}: Date file created
}
\usage{
spatial_tract10_to_hra20_region20_geog
}
\description{
The are based on the fractional overlap approach in spatagg::create_xwalk and broadly
follow that file structure.
}
\keyword{datasets}
