% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_tract10_to_puma}
\alias{spatial_tract10_to_puma}
\title{Census 2010 crosswalk tracts to PUMAS}
\format{
A data.table with 398 rows and 5 variables:
\code{wa}, \code{kingco}, \code{tract10}, \code{puma}, \code{uploaded}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_tract10_to_puma
}
\description{
Census 2010 crosswalk tracts to PUMAS
}
\keyword{datasets}
