% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_tract20_to_hra20_geog}
\alias{spatial_tract20_to_hra20_geog}
\title{Spatial crosswalk between 2020 census tracts and 2020 HRAs. There is
one row between each tract and HRA that intersects geographically}
\format{
A data.table with 1,083 rows and 11 variables:

\code{GEOID20}: tract id,

\code{hra20_id}: HRA numeric id,

\code{s2t_fraction}: Percent of the area in tract that is also in hra20_id,

\code{isect_amount}: Number of square survey feet in the intersection of tract and hra20_id,

\code{tcoverage_amount}: Area of the hra20_id in any tract code,

\code{target_amount}: Total area of the HRA (including areas not covered by tracts),

\code{hra_name}: Name of the HRA,

\code{source_id}: tract,

\code{target_id}: hra20_id,

\code{method}: method to produce intersections,

\code{creation_date}: Date file created
}
\source{
2020 Census tracts (\verb{//dphcifs/APDE-CDIP/Shapefiles/Census_2020/tract/}) intersected
with the 2020 HRA shapefile (geographic overlap method).
}
\usage{
spatial_tract20_to_hra20_geog
}
\description{
These are based on the fractional overlap approach in \code{\link[spatagg:create_xwalk]{spatagg::create_xwalk()}} and broadly
follow that file structure.
}
\keyword{datasets}
