% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_tract_to_county_council_dist}
\alias{spatial_tract_to_county_council_dist}
\title{Crosswalk between 2010 census tracts and King County Council Districts}
\format{
A data.table with 538 rows and 32 variables: \code{tract},
\code{kccdst_assigned}, \code{fid}, \code{objectid}, \code{fid_kccdst},
\code{kccdst}, \code{councilmem}, \code{fid_tracts}, \code{geo_id_trt},
\code{feature_id}, \code{tract_lbl}, \code{tract_str}, \code{tract_int},
\code{tract_flt}, \code{tract_del}, \code{trtlabel_f}, \code{trtlabel_c},
\code{trtlabel_t}, \code{county_str}, \code{county_int}, \code{state_str},
\code{state_int}, \code{level_1}, \code{level_2}, \code{level_3},
\code{tract_area}, \code{tract_peri}, \code{logrecno17}, \code{shape_area},
\code{shape_leng}, \code{shape_ar_1}, \code{tract_perc}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_tract_to_county_council_dist
}
\description{
Crosswalk between 2010 census tracts and King County Council Districts
}
\keyword{datasets}
