% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_zip_admin}
\alias{spatial_zip_admin}
\title{Crosswalk zip codes to cities / towns}
\format{
A data.table with 122 rows and 3 variables:
\code{zip}, \code{zip_type},\code{city}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_zip_admin
}
\description{
Crosswalk zip codes to cities / towns
}
\keyword{datasets}
