% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_zip_city_region_vid}
\alias{spatial_zip_city_region_vid}
\title{Crosswalk zip codes to cities and regions}
\format{
A data.table with 120 rows and 4 variables:
  \code{zip}, \code{city}, \code{region}, \code{region_vid}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_zip_city_region_vid
}
\description{
Crosswalk zip codes to cities and regions
}
\keyword{datasets}
