% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_zip_hca}
\alias{spatial_zip_hca}
\title{Crosswalk FOR HCA MEDICAID KC definition ... zip >> city}
\format{
A data.table with 133 rows and 5 variables:
\code{zip}, \code{zip_type}, \code{city}, \code{primary_city}, \code{two_counties}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_zip_hca
}
\description{
133 zip codes, with zip type, city, primary_city, used to define King County
in Health Care Authority Medicaid data extracts. Includes 98354 and 98422,
which barely cross into south KC). ZIPs spanning county lines are noted in
the geo_multi_county field.
}
\keyword{datasets}
