% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_zip_to_city_council_dist}
\alias{spatial_zip_to_city_council_dist}
\title{Crosswalk zip code to Seattle city council districts}
\format{
A data.table with 24 rows and 2 variables:
\code{zip}, \code{district}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_zip_to_city_council_dist
}
\description{
Crosswalk zip code to Seattle city council districts
}
\keyword{datasets}
