% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_zip_to_region}
\alias{spatial_zip_to_region}
\title{Crosswalk zip codes to regions}
\format{
A data.table with 120 rows and 6 variables:
  \code{zip}, \code{city}, \code{cc_region}, \code{po_box},
  \code{office_building}, \code{notes}.
}
\source{
\url{https://www.mysterydata.com}
}
\usage{
spatial_zip_to_region
}
\description{
Crosswalk zip codes to regions
}
\keyword{datasets}
