% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_chars}
\alias{synthetic_chars}
\title{Synthetic CHARS Data for Testing and Development}
\format{
A data.table with 8 columns and more than 250,000 rows:
\itemize{
\item \strong{\code{seq_no}}: Unique patient-visit level identifier (one per row)
\item \strong{\code{diag1}}: ICD-10-CM Primary diagnosis code
\item \strong{\code{injury_nature_broad}}: injuries are defined by searching all available diagnosis fields on the
hospital discharge record (there can be a maximum of 54 diagnosis fields
in CHARS data)
\item \strong{\code{injury_nature_narrow}}: injuries are defined by following CDC's recommended 'narrow' approach, which
requires that the principal diagnosis of an injury hospitalization be a
nature-of-injury ICD10-CM code
\item \strong{\code{injury_intent}}: Injury intent classification, i.e., 'assault', 'intentional',
'legal', 'undetermined', 'unintentional'.
\item \strong{\code{injury_mechanism}}: Injury mechanism classification, e.g., 'bites_stings',
'drowning', 'fall', 'firearm', 'overexertion, 'poisoning', etc.
\item \strong{\code{temperament}}: A categorical 'demographic' indicator with three levels:
'Active', 'Moderate', and 'Calm'.
\item \strong{\code{creation_date}}: A simple time stamp of the date the table was updated
}
}
\source{
Generated by \code{\link[rads.internal:make_synthetic_chars]{rads.internal::make_synthetic_chars()}}.
}
\usage{
synthetic_chars
}
\description{
A synthetic dataset derived from real Comprehensive Hospital Abstract
Reporting System (CHARS) data for use in testing and development. This
privacy-safe dataset contains injury-related diagnosis data with an
arbitrary categorical descriptor (temperament).
}
\examples{
library(data.table)

# Load the synthetic CHARS data
data(synthetic_chars)

# View structure
str(synthetic_chars)

# Check the distribution of temperament
table(synthetic_chars$temperament)

}
\keyword{datasets}
