% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_infdeath}
\alias{synthetic_infdeath}
\title{Synthetic Infant Death Data for Testing and Development}
\format{
A data.table with 4 columns and more than 2,000 rows:
\itemize{
\item \strong{\code{year}}: Year of death
\item \strong{\code{sex}}: Infant's sex
\item \strong{\code{underlying_cod_code}}: ICD-10 underlying cause of death code, cleaned via \code{\link[rads:death_icd10_clean]{rads::death_icd10_clean()}}
\item \strong{\code{creation_date}}: A simple time stamp of the date the table was updated
}
}
\source{
Generated by \code{\link[rads.internal:make_synthetic_infdeath]{rads.internal::make_synthetic_infdeath()}}.
}
\usage{
synthetic_infdeath
}
\description{
A synthetic dataset derived from real death data for use in testing
and development.
}
\examples{
library(data.table)

# Load the synthetic CHARS data
data(synthetic_infdeath)

# View structure
str(synthetic_infdeath)

# Check the distribution of sex
table(synthetic_infdeath$sex)

}
\keyword{datasets}
