% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_population}
\alias{synthetic_population}
\title{Synthetic Population Data for Testing and Development}
\format{
A data.table with 6 columns and more than 14,000 rows:
\itemize{
\item \strong{\code{year}}: Calendar year of population estimate
\item \strong{\code{gender}}: 'Female' or 'Male'
\item \strong{\code{race}}: 'AIAN', 'Asian', 'Black', 'Hispanic', 'Multiple race', 'NHPI', or 'White
\item \strong{\code{age}}: Age
\item \strong{\code{pop}}: Population estimate
\item \strong{\code{creation_date}}: A simple time stamp of the date the table was updated
}
}
\source{
Generated by \code{\link[rads.internal:make_synthetic_population]{rads.internal::make_synthetic_population()}}.
}
\usage{
synthetic_population
}
\description{
A 'synthetic' population dataset derived \href{https://ofm.wa.gov/}{WA OFM} estimates for use in testing
and development. Estimates are current as of the \code{creation_date} in the dataset.
}
\examples{
library(data.table)

# Load the synthetic CHARS data
data(synthetic_population)

# View structure
str(synthetic_population)

# Population of Females in Washington State in 2022
sum(synthetic_population[gender == 'Female' &
                         year == 2022]$pop)

}
\keyword{datasets}
