% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_death_injury_matrix}
\alias{icd10_death_injury_matrix}
\title{ICD10 Death Injury Matrix}
\format{
a 'long' data.table with 1,398 rows and four columns: \code{intent}, \code{mechanism}, \code{orig.coding}, & \code{icd10}.
}
\source{
WA DOH CHAT: \url{https://secureaccess.wa.gov/doh/chat/Content/FilesForDownload/CodeSetDefinitions/CHATInjury(ICE)codes.pdf}.
Converted to XLSX with Adobe Acrobat and then used Excel to convert to CSV.
}
\usage{
icd10_death_injury_matrix
}
\description{
ICD10 Death Injury Matrix (ICE: International Collaborative Effort on Injury Statistics)
}
\note{
This function uses ICD10 mortality codes and should not be used with  ICD10-CM hospitalization data.
}
\references{
\url{https://www.flhealthcharts.gov/ChartsReports/_SupportFiles/InjuryDeathProfileReference.pdf}
}
\keyword{datasets}
