% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_nchs113causes}
\alias{icd_nchs113causes}
\title{NCHS 113 Selected Causes of Death (long).}
\format{
A data.table with 2975 rows and three variables: \code{causeid}, \code{cause.of.deaths},
  \code{icd10}. Each row maps one ICD 10 code to one of the 113 causes of death.
}
\source{
\url{https://www.cdc.gov/nchs/data/dvs/Part9InstructionManual2020-508.pdf}
}
\usage{
icd_nchs113causes
}
\description{
NCHS 113 Selected Causes of Death (long). Cause id 95 ("All other diseases (Residual)") uses codings discovered on a PHSKC
    network drive. They were propbably created by Mike Smyzer, who has retired. These
    are being used instead of the CDC codes due to better alignment with WA DOH. The
    official CDC coding is now saved as cause id 114 "CDC version of cause id 95 (Residual)".
}
\note{
Per CDC guidance, our cause list contains COVID-19 (U07.1) in cause id 17 ("Other and unspecified
    infectious and parasitic diseases and their sequelae")
}
\references{
https://secureaccess.wa.gov/doh/chat/Content/FilesForDownload/CodeSetDefinitions/NCHS113CausesOfDeath.pdf
}
\keyword{datasets}
