% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_nchs113causes_raw}
\alias{icd_nchs113causes_raw}
\title{NCHS 113 Selected Causes of Death (raw)}
\format{
A data.table with 114 rows and 11 variables: \code{causeid},\code{cause.of.death},\code{icd10},
\code{cause.category}, \code{leading.cause.group.num},\code{leading.cause.group},\code{leading.cause.group.alias},\code{level1},
\code{level2},\code{level3}, & \code{level4}. Each of the first 113 rows is for one of the 113 NCHS causes of death,
with the \code{icd10} column summarizing all the relevant codes for the given cause of death.
}
\source{
Downloaded from \url{https://ibis.health.utah.gov/ibisph-view/query/NCHS113.html} on 9/22/2021. It is a machine readable version of
\url{https://secureaccess.wa.gov/doh/chat/Content/FilesForDownload/CodeSetDefinitions/NCHS113CausesOfDeath.pdf}
}
\usage{
icd_nchs113causes_raw
}
\description{
NCHS 113 Selected Causes of Death (raw). Used to create \code{icd_nchs113causes}
with \code{/data-raw/icd_nchs113causes.R}
}
\note{
Per CDC guidance, our cause list contains COVID-19 (U07.1) in cause id 17 ("Other and unspecified
    infectious and parasitic diseases and their sequelae")

    Cause id 95 ("All other diseases (Residual)") uses codings discovered on a PHSKC
    network drive. They were probably created by Mike Smyser, who has retired. These
    are being used instead of the CDC codes due to better alignment with WA DOH.

    The official CDC coding for Cause id 95 is now saved as cause id 114 "CDC version of cause id 95 (Residual)".
}
\references{
\url{https://www.cdc.gov/nchs/data/dvs/Part9InstructionManual2020-508.pdf}
}
\keyword{datasets}
