% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_nchs130causes_raw}
\alias{icd_nchs130causes_raw}
\title{NCHS 130 Selected Causes of Infant Death (raw)}
\format{
A data.table with 130 rows and 11 variables: \code{causeid},\code{cause.of.death},\code{icd10},
\code{cause.category}, \code{leading.cause.group.num},\code{leading.cause.group},\code{leading.cause.group.alias},\code{level1},
\code{level2},\code{level3}, & \code{level4}. Each row is for one of the 130 NCHS causes of infant death,
with the \code{icd10} column summarizing all the relevant codes for the given cause of death.
}
\source{
Downloaded from
\url{https://secureaccess.wa.gov/doh/chat/Content/FilesForDownload/TechnicalNotes.pdf}
on 2/15/2023. It should represent table in
\url{https://www.cdc.gov/nchs/data/dvs/Part9InstructionManual2020-508.pdf}, on pages 18-20.
}
\usage{
icd_nchs130causes_raw
}
\description{
NCHS 130 Selected Causes of Infant Death (raw). Used to create \code{icd_nchs130causes}
with \code{/data-raw/icd_nchs130causes.R}
}
\note{
This table is based on a download from WA DOH (see sources below). The PDF was exported to an Excel file and the Excel
file was manually cleaned to make it machine readable. The `Levels` based on the hierarchy shown in the CHAT infant
mortality module's 'NCHS130 Groupings' code set. The `leading.cause.group` information is from the CHAT infant mortality
module's 'Leading Infant Causes' code set. The `leading.cause.group.alias` are from Danny's imagination and should not be assumed
to match any standard.
}
\references{
\url{https://www.cdc.gov/nchs/data/dvs/Part9InstructionManual2020-508.pdf}
}
\keyword{datasets}
