% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_other_causes_of_death}
\alias{icd_other_causes_of_death}
\title{Other Causes of Death (long)}
\format{
A data.table with 429 rows and three variables: \code{cause.of.deaths},
  \code{orig.coding}, and \code{icd10}. Each row maps one specific ICD 10
  code to a given cause of death.
}
\source{
Drug-induced and drug-overdose are extracted from
\url{https://www.cdc.gov/nchs/data/nvsr/nvsr70/nvsr70-08-508.pdf}.
}
\usage{
icd_other_causes_of_death
}
\description{
Other Causes of Death (long).
}
\details{
Tidied long version of \code{/data-raw/icd_other_causes_of_death_raw}.
Converted by \code{/data-raw/icd_other_causes_of_death.R}.
}
\note{
This table contains ICD 10 definitions of causes of death that are of
      interest for systematic analyses, but are not included in the CDC 113
      Causes of Death (\code{/data/icd_nchs113causes.rda}) or the injury
      matrix (\code{/data/icd10_death_injury_matrix.rda})
}
\references{
\url{}
}
\keyword{datasets}
