% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd_other_causes_of_death}
\alias{icd_other_causes_of_death}
\title{Other Causes of Death (long)}
\format{
A data.table with 1,496 rows and four variables: \code{cause.of.deaths},
  \code{orig.coding}, \code{icd10}, and \code{source}. Each row maps one
  specific ICD 10 code to a given cause of death.
}
\source{
CDC PDF:
\url{https://www.cdc.gov/nchs/data/nvsr/nvsr70/nvsr70-08-508.pdf}.

DOH CHAT (User Guides >> Definition of "Special Code Sets")
\url{https://secureaccess.wa.gov/doh/chat/Entry.mvc}
}
\usage{
icd_other_causes_of_death
}
\description{
This table contains ICD 10 definitions of causes of death that are of
interest for systematic analyses, but are not included in the CDC 113
Causes of Death (\code{/data/icd_nchs113causes.rda}) or the injury
matrix (\code{/data/icd10_death_injury_matrix.rda}). It includes causes such
as carbon monoxide deaths, drug related deaths, heat stress deaths, etc.
}
\note{
Tidied long version of \code{/data-raw/icd_other_causes_of_death_raw} PLUS
an Excel workbook downloaded from DOH/CHAT and saved to SharePoint.
Created by \code{/data-raw/icd_other_causes_of_death.R}.
}
\references{
\url{https://www.cdc.gov/nchs/data/nvsr/nvsr70/nvsr70-08-508.pdf}
}
\keyword{datasets}
