% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc_chi_byvars}
\alias{misc_chi_byvars}
\title{The definitive reference table for CHI cat/varname/group/group_alias values}
\format{
a data.table with a flexible number of rows and six columns:
\code{cat}, \code{varname}, \code{group}, \code{keepme},
\code{notes} & \code{creation_date}.
}
\source{
A version controlled copy of the data in SharePoint >>
Community Health Indicators >> CHI-Vizes >> CHI-Standards-TableauReady Output.xlsx
}
\usage{
misc_chi_byvars
}
\description{
The definitive reference table for CHI cat/varname/group/group_alias values.
Used in CHI analyses as the by-variables (cat1 & cat1 in CHI parlance).
}
\note{
This replaces any and all old standard found in documentation or in
existing SQL databases.
}
\keyword{datasets}
