% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation_soc_2010_definitions}
\alias{occupation_soc_2010_definitions}
\title{2010 BLS Standard Occupational Classification (SOC) definitions.}
\format{
A data.table with 840 rows and four variables: \code{'SOC.Code'}, \code{'SOC.Title'},
\code{'SOC.Definition'}, \code{'uploaded'}.
}
\source{
\url{https://www.bls.gov/soc/soc_2010_definitions.xls}
}
\usage{
occupation_soc_2010_definitions
}
\description{
2010 BLS Standard Occupational Classification (SOC) definitions.
}
\details{
Detailed definitions for each 2010 BLS SOC code / title
}
\keyword{datasets}
