% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_block20_to_hra20_to_region20}
\alias{spatial_block20_to_hra20_to_region20}
\title{Spatial crosswalk for Census 2020 blocks to 2020 HRAs and 2020 regions}
\format{
A data.table with 27,686 rows and 6 variables:

\code{hra20_id}: Numeric id for the HRA,

\code{GEOID20}: 2020 Census block id,

\code{hra20_name}: Name of the HRA,

\code{region_id}: numeric id for the region,

\code{region_name}: Name of the region,

\code{creation_date}: Date the file was created
}
\usage{
spatial_block20_to_hra20_to_region20
}
\description{
Spatial crosswalk for Census 2020 blocks to 2020 HRAs and 2020 regions
}
\keyword{datasets}
