% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_block20_to_hra20_to_region20}
\alias{spatial_block20_to_hra20_to_region20}
\title{Spatial crosswalk for Census 2010 blocks to 2020 HRAs and 2020 regions}
\format{
A data.table with 27686 rows and 6 variables:
  \code{hra_id}: Numeric id for the HRA,
  \code{GEOID20}: 2020 Census block id ,
  \code{hra_name}: Name of the HRA,
  \code{reg_id}: numeric id for the region,
  \code{reg_nm}: Name of the region
  \code{creation}: Date the file was created
}
\usage{
spatial_block20_to_hra20_to_region20
}
\description{
Spatial crosswalk for Census 2010 blocks to 2020 HRAs and 2020 regions
}
\keyword{datasets}
