% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_geoid20_to_hra20_acs}
\alias{spatial_geoid20_to_hra20_acs}
\title{Spatial crosswalk for ACS Census 2020 geographies to 2020 HRAs}
\format{
A data.table with 547 rows and 6 variables:

\code{hra20_id}: Numeric id for the 2020 HRA,

\code{hra20_name}: Official name for the 2020 HRA,

\code{place20}: place 2020 GEOID,

\code{tract20}: tract 2020 GEOID,

\code{bg20}: bg 2020 GEOID,

\code{creation}: Date the file was created
}
\usage{
spatial_geoid20_to_hra20_acs
}
\description{
Different American Community Survey (ACS) estimates are released
at a variety of geographic levels: block groups, tracts, places (cities,
towns, and Census Designated Places (CDPs)), etc. Since HRAs are properly
defined by blocks and block level estimates are not available, this
crosswalk represents a best attempt to create a crosswalk using the largest
possible sub-HRA geographies so as to prevent unnecessary inflation in the
standard error during calculation of ACS HRA and Region level estimates.
}
\keyword{datasets}
