% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_geoid20_to_sccd13_acs}
\alias{spatial_geoid20_to_sccd13_acs}
\title{Spatial crosswalk for ACS Census 2020 geographies to 2013 Seattle City Council Districts}
\format{
A data.table with 177 rows and 6 variables:
  \code{sccd13_id}: Seattle City Council district number,
  \code{sccd13_name}: Seattle City Council district name,
  \code{place20}: place 2020 GEOID,
  \code{tract20}: tract 2020 GEOID,
  \code{bg20}: bg 2020 GEOID,
  \code{creation}: Date the file was created
}
\usage{
spatial_geoid20_to_sccd13_acs
}
\description{
Different American Community Survey (ACS) estimates are released
at a variety of geographic levels: block groups, tracts, places (cities,
towns, and Census Designated Places (CDPs)), etc. This
crosswalk represents a best attempt to create a crosswalk using the largest
possible sub-district geographies so as to prevent unnecessary inflation in the
standard error during calculation of ACS Seattle City Council level estimates.
}
\keyword{datasets}
