% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_hra20_to_region20}
\alias{spatial_hra20_to_region20}
\title{Spatial crosswalk between 2020 HRAs, and 2020 regions.}
\format{
A data.table with 666 rows and 111 variables:
  \code{hra_id}: Numeric id for the HRA,
  \code{hra_name}: Name of the HRA,
  \code{reg_id}: numeric id for the region,
  \code{reg_nm}: Name of the region
}
\usage{
spatial_hra20_to_region20
}
\description{
Spatial crosswalk between 2020 HRAs, and 2020 regions.
}
\keyword{datasets}
