% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_crosswalks.R
\docType{data}
\name{spatial_ids_and_names}
\alias{spatial_ids_and_names}
\title{Crosswalk between geo_id and geo_names for an assortment of gegraphies}
\format{
A data.table with 506 rows and 6 variables:

  \code{geo_id_code}: Numeric code of the geography,

  \code{geo_id}: Name of the geography,

  \code{type}: Geography type
}
\usage{
spatial_ids_and_names
}
\description{
Mostly a helper file for rads::get_population
}
\keyword{datasets}
