% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_school_codes_to_names}
\alias{spatial_school_codes_to_names}
\title{Crosswalk between 2010 school district id, name, state id, county code, and esd name}
\format{
A data.table with 336 rows and 7 variables: \code{scd_id}, \code{geo_year},
  \code{scd_name}, \code{scd_state_id}, \code{scd_cou_code}, \code{scd_lea_code},
  \code{scd_esd_name}.
}
\source{
kcitazrhpasqlprp16.azds.kingcounty.gov [ref].[pop_scd_crosswalk]
}
\usage{
spatial_school_codes_to_names
}
\description{
Crosswalk between 2010 school district id, name, state id, county code, and esd name
}
\keyword{datasets}
