% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_tract20_to_hra20_pop}
\alias{spatial_tract20_to_hra20_pop}
\title{Spatial crosswalk between 2020 census tracts and 2020 HRAs. There is
one row between each tract and HRA that shares population.}
\format{
A data.table with 666 rows and 111 variables:

  \code{GEOID20}: Census 2020 id,

  \code{hra20_id}: HRA numeric id,

  \code{s2t_fraction}: Percent of the population in tract that is also in hra20_id,

  \code{isect_amount}: # of people who live in the intersection of tract and hra20_id,

  \code{tcoverage_amount}: Population of the hra20_id in any tract code,

  \code{target_amount}: Total number of people within the HRA (including areas not covered by tracts),

  \code{hra_name}: Name of the HRA,

  \code{source_id}: tract,

  \code{target_id}: hra20_id,

  \code{method}: method to produce intersections

  \code{creation_date}: Date file created
}
\usage{
spatial_tract20_to_hra20_pop
}
\description{
The are based on the point pop approach in spatagg::create_xwalk and broadly
follow that file structure.
}
\keyword{datasets}
