% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_zip_city_region_scc}
\alias{spatial_zip_city_region_scc}
\title{Crosswalk zip codes >> cities >> regions (& SCC)}
\format{
A data.table with 126 rows and 7 columns:
\code{zip}, \code{zip_type}, \code{city}, \code{city_primary}, \code{multi_county},
\code{office_building}, \code{notes}.
}
\source{
combined spatial_zip_admin, spatial_zip_to_city_council_dist,
spatial_zip_to_region, & spatial_zip_city_region_vid & then deleted the
originals
}
\usage{
spatial_zip_city_region_scc
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Crosswalk zip codes >> cities >> regions (& SCC)
}
\details{
\strong{DEPRECATED} beginning version 1.0.19

Version 1.0.19 removed \code{region}, \code{region_id}, \code{region_vid}, & \code{scc}.
}
\keyword{datasets}
