% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_zip_city_region_scc}
\alias{spatial_zip_city_region_scc}
\title{Crosswalk zip codes >> cities >> regions (& SCC)}
\format{
A data.table with 126 rows and 11 variables:
\code{zip}, \code{zip_type}, \code{city}, \code{city_primary}, \code{region},
\code{region_id}, \code{region_vid}, \code{scc}, \code{multi_county},
\code{office_building}, \code{notes}.
}
\source{
combined spatial_zip_admin, spatial_zip_to_city_council_dist,
spatial_zip_to_region, & spatial_zip_city_region_vid & then deleted the
originals
}
\usage{
spatial_zip_city_region_scc
}
\description{
Crosswalk zip codes >> cities >> regions (& SCC)
}
\keyword{datasets}
