% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hra2020_update.R
\docType{data}
\name{spatial_zip_to_hra20_pop}
\alias{spatial_zip_to_hra20_pop}
\title{Spatial crosswalk between ZIP codes and  2020 HRAs}
\format{
A data.table with 233 rows and 11 variables:
  \code{ZIP}: ZIP code,
  \code{hra20_id}: HRA numeric id,
  \code{s2t_fraction}: Percent of the population in ZIP that is also in hra20_id,
  \code{isect_amount}: # of people who live in the intersection of ZIP and hra20_id,
  \code{tcoverage_amount}: Population of the hra20_id in any ZIP code,
  \code{target_amount}: Total number of people within the HRA (including areas not covered by ZIPs)
  \code{hra_name}: Name of the HRA,
  \code{source_id}: ZIP,
  \code{target_id}: hra20_id,
  \code{method}: method to produce intersections
  \code{creation_date}: Date file created
}
\usage{
spatial_zip_to_hra20_pop
}
\description{
There is one row between each ZIP and HRA that shares population.
}
\details{
The are based on the point pop approach in spatagg::create_xwalk and broadly
follow that file structure.

Note: ZIP codes are not completely nested within King County. As such, end users
may want to further process this file to handle edge effects
}
\keyword{datasets}
