% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_birth}
\alias{synthetic_birth}
\title{Synthetic Birth Data for Testing and Development}
\format{
A data.table with 11 columns and more than 70,000 rows:
\itemize{
\item \strong{\code{chi_year}}: Birth year, randomly reassigned while preserving the
empirical distribution
\item \strong{\code{chi_age}}: Maternal age, randomly reassigned while preserving the
empirical distribution
\item \strong{\code{chi_geo_kc}}: A King County, WA specific indicator, randomly assigned
as \code{"King County"} or \code{NA}
\item \strong{\code{chi_sex}}: Infant sex, randomly reassigned while preserving the
empirical distribution
\item \strong{\code{race4}}: OMB race/ethnicity, with Hispanic as a race, randomly
reassigned while preserving the empirical distribution
\item \strong{\code{birth_weight_grams}}: The recorded birth weight in grams
\item \strong{\code{preterm}}: Binary flag for preterm delivery
\item \strong{\code{fetal_pres}}: Fetal presentation (Breech, Cephalic, or Other)
\item \strong{\code{kotelchuck}}: Binary flag for early and adequate prenatal care, as
measured by the Kotelchuck index
\item \strong{\code{mother_birthplace_state}}: Mother's birthplace state, randomly
reassigned while preserving the empirical distribution
\item \strong{\code{creation_date}}: A simple time stamp of the date the table was updated
}
}
\source{
Generated by \code{\link[rads.internal:make_synthetic_birth]{rads.internal::make_synthetic_birth()}}.
}
\usage{
synthetic_birth
}
\description{
A synthetic dataset derived from real 2019-2023 birth data for use in
testing and development. This privacy-safe dataset preserves the
distributions of key demographic and clinical birth characteristics.
}
\examples{
library(data.table)

# Load the synthetic birth data
data(synthetic_birth)

# View structure
str(synthetic_birth)

# Check the distribution of infant sex
table(synthetic_birth$chi_sex)

}
\keyword{datasets}
