% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_death}
\alias{synthetic_death}
\title{Synthetic Death Data for Testing and Development}
\format{
A data.table with 27 columns and more than 30,000 rows:
\itemize{
\item \strong{\code{underlying_cod_code}}: ICD-10 underlying cause of death code, cleaned
via \code{\link[rads:death_icd10_clean]{rads::death_icd10_clean()}}
\item \strong{\code{record_axis_code_1}} through \strong{\code{record_axis_code_20}}: ICD-10
contributing causes of death codes, cleaned via \code{\link[rads:death_icd10_clean]{rads::death_icd10_clean()}}
\item \strong{\code{date_of_birth}}: Date of birth (class Date)
\item \strong{\code{date_of_death}}: Date of death (class Date)
\item \strong{\code{chi_geo_kc}}: A King County, WA specific indicator, randomly assigned
as \code{"King County"} or \code{NA}
\item \strong{\code{temperament}}: A categorical 'demographic' indicator with three levels:
\code{'Active'}, \code{'Moderate'}, and \code{'Calm'}
\item \strong{\code{chi_age}}: Age at death in integer years
\item \strong{\code{creation_date}}: A simple time stamp of the date the table was updated
}
}
\source{
Generated by \code{\link[rads.internal:make_synthetic_death]{rads.internal::make_synthetic_death()}}.
}
\usage{
synthetic_death
}
\description{
A synthetic dataset derived from real death data for use in testing
and development. This privacy-safe dataset contains cause of death (ICD-10),
date of birth, and and date of death data with an arbitrary categorical
descriptor (temperament).
}
\examples{
library(data.table)

# Load the synthetic CHARS data
data(synthetic_death)

# View structure
str(synthetic_death)

# Check the distribution of temperament
table(synthetic_death$temperament)

}
\keyword{datasets}
